Basic framework port with new action abstraction.
The action abstraction was suggessted by linus@mullvad.net. It allows
the core of the framework to be agnostic to the required async and
timers used to schedule actions to pad and block. The framework, after
setup with zero or more machines, now maps zero or more events to a
number of actions (at most, one action per machine) at a given point
in time. This has the added benefit of being able to handle the case
when the user of the framework gets overloaded with many buffered events
inbetween calls to the framework.

Still have to port a number of tests and implement a PoC of using the
framework with the new abstraction.