Fundamental change to State struct, removed a dist.
Earlier, State had two distributions: block and size. Block was used if
a block action was to be done and size for a padding action. This is now
replaced with a single action distribution. What the distribution is to
be used for depends on a new flag, actions_is_block, which defaults to
false (so default is to pad). This saves a significant number of bytes
per state and improves readability.