Fns to get num machines and create a MachineId.
The number of machines is a helper to learn how many machines are
running in an instance of the framework. The helper may be useful in
concjunction with the ability to create MachineIds for testing purposes.
Normal usage should use the MachineIds returned in Actions by the
framework.

There are two downsides with the current approach:
- We use a usize internally for a MachineId. This makes it
  straightforward to use without any conversions in the framework, but
  the number of bytes used by usize depends on platform/arch, so
  serialization can be unnecessarily complicated.
- If a user of the framework ever triggers an event with a MachineId
  that does not exist in the framework, then this does not raise any
  error nor panics. This can make it harder to detect the wrong usage of
  the framework.

We might want to makes changes here later due to the above, but let's do
that when we get there.

This resolves #3 for now.