Semantics change wrt replacing packets and length.
Earlier, replaced packets had to be of the same length as the padding
packet being replaced. This made it every hard to create constant-rate
defenses for transports with variable length packets. Now, instead, the
packet MAY be of the same length. This turns it into an implementation
question. For now, I think it's reasonable to just ignore packet sizes.