s/nonpadding/normal
We stick with padding instead of say chaff or cover, because chaff and
cover are goals/reasons for padding. Padding is also strongly associated
with Adaptive Padding and Tor's Circuit Padding Framework, the roots of
Maybenot.

We replace "non-padding" with "normal" to make it more distinct when,
e.g., grepping logs for padding or normal events.