Events {Padding,Normal,Tunnel}+{Sent,Recv}.
Another event refactor and addition: removed the newly added
NormalQueued and PaddingQueued events, and introduced the TunnelSent and
TunnelRecv events instead. Now, we have {Padding,Normal,Tunnel}+{Sent,Recv}
events. Every event acts on a packet, and a packet is first either a
normal or padding packet, that is then sent through the tunnel we're
protecting. Once a packet enters (or exits) the tunnel, nobody should be
able to distinguish between normal and padding packets. This is only
possible before encryption/after decryption. How and where to
distinguish between normal, padding, and tunnel packets depends on where
Maybenot is integrated.